Deprecate and refactor IMenuItemClassNames to .types for export
