+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenu.classNames.ts
@@ -3,22 +3,6 @@ import { ITheme, mergeStyleSets } from '../../Styling';
 import { IVerticalDividerClassNames } from '../Divider/VerticalDivider.types';
 import { memoizeFunction } from '../../Utilities';
 
-export interface IMenuItemClassNames {
-  item: string;
-  divider: string;
-  root: string;
-  linkContent: string;
-  icon: string;
-  checkmarkIcon: string;
-  subMenuIcon: string;
-  label: string;
-  secondaryText: string;
-  splitContainer: string;
-  splitPrimary: string;
-  splitMenu: string;
-  linkContentMenu: string;
-}
-
 export const getSplitButtonVerticalDividerClassNames = memoizeFunction(
   (theme: ITheme): IVerticalDividerClassNames => {
     return mergeStyleSets(getDividerClassNames(theme), {
