+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenuItemWrapper/ContextualMenuButton.test.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
-import { IContextualMenuItem } from '../ContextualMenuItem.types';
+import { IContextualMenuItem, IMenuItemClassNames } from '../ContextualMenuItem.types';
 import { ContextualMenuButton } from './ContextualMenuButton';
-import { IMenuItemClassNames } from 'office-ui-fabric-react/lib/components/ContextualMenu/ContextualMenu.classNames';
 
 describe('ContextualMenuButton', () => {
   describe('creates a normal button', () => {
